+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 * Infer parent ID mode from the request when retrieving the parent (target) page to avoid `notfound`.
+* Log the actual REST API error message and not the one meant for the user. 
 
 ## 3.54.0 (2023-08-16)